Refactor folders structure
Extract components to components folder to implement request changes from
 review
